+++ src/components/admin/MpAdminMenu.vue
@@ -2,54 +2,52 @@
 import { defineAsyncComponent, onMounted, ref } from 'vue'
 import { useConfirm } from 'primevue/useconfirm'
 import { useToast } from 'primevue/usetoast'
-import { useCarouselStore } from '@/store/carousel.js'
+import { useMenuStore } from '@/store/menu.js'
 
 const MpAdminModal = defineAsyncComponent(/* webpackChunkName: "mpAdminModal" */() => import('@/components/admin/MpAdminModal.vue'))
 
-const carouselStore = useCarouselStore()
+const menuStore = useMenuStore()
 const confirm = useConfirm()
 const toast = useToast()
 
-const openModalCarousel = ref(false)
+const openModalMenu = ref(false)
 
 const manageValueModal = (value) => {
-  openModalCarousel.value = value
+  openModalMenu.value = value
 }
 
 const configModal = ref({})
 
-const manageModalCarousel = (action, type, data = null) => {
+const manageModalMenu = (action, type, data = null) => {
   if (action === 'add') {
     configModal.value = {
-      header: 'Agregar imagen en el carrusel',
-      description: 'Completa los campos para agregar una nueva imagen en el carrusel.',
+      header: 'Agregar menú',
+      description: 'Completa los campos para agregar un nuevo menú.',
       acceptButton: 'Agregar',
       cancelButton: 'Cancelar',
       type,
       action,
-      data,
-      containImage: true
+      data
     }
   }
 
   if (action === 'edit') {
     configModal.value = {
-      header: 'Editar imagen del carrusel',
-      description: 'Completa los campos para editar la imagen del carrusel.',
+      header: 'Editar menú',
+      description: 'Completa los campos para editar el menú seleccionado.',
       acceptButton: 'Editar',
       cancelButton: 'Cancelar',
       type,
       action,
-      data,
-      containImage: true
+      data
     }
   }
-  openModalCarousel.value = true
+  openModalMenu.value = true
 }
 
-const confirmDeleteCarousel = (id, image) => {
+const confirmDeleteMenu = (id) => {
   confirm.require({
-    message: '¿Quieres eliminar esta imagen del carrusel?',
+    message: '¿Quieres eliminar este menú?',
     header: 'Confirmar',
     icon: 'pi pi-info-circle mr-2',
     rejectLabel: 'Cancelar',
@@ -57,8 +55,8 @@ const confirmDeleteCarousel = (id, image) => {
     rejectClass: 'transparent',
     acceptClass: 'bg-danger text-white',
     accept: () => {
-      toast.add({ severity: 'info', summary: 'Eliminado', detail: 'La imagen del carrusel ha sido eliminada.', life: 3000 })
-      carouselStore.deleteCarousel(id, image)
+      toast.add({ severity: 'info', summary: 'Eliminado', detail: 'La menú ha sido eliminado.', life: 3000 })
+      menuStore.deleteMenu(id)
     },
     reject: () => {
       toast.add({ severity: 'error', summary: 'Cancelado', detail: 'Se ha cancelado la operación.', life: 3000 })
@@ -67,58 +65,50 @@ const confirmDeleteCarousel = (id, image) => {
 }
 
 onMounted(() => {
-  carouselStore.getCarousel()
+  menuStore.getMenu()
 })
 </script>
 
 <template>
   <MpAdminModal
-    :visible="openModalCarousel"
+    :visible="openModalMenu"
     :configModal="configModal"
     @manageModal="manageValueModal"
   />
-  <div v-if="carouselStore.isLoading" class="container mx-auto grid gap-4 mt-1">
+  <div v-if="menuStore.isLoading" class="container mx-auto grid gap-4 mt-1">
     <div class="flex flex-wrap align-items-center justify-content-between gap-2">
-      <span class="text-xl text-900 font-bold">Carrusel de imágenes</span>
+      <span class="text-xl text-900 font-bold">Menú</span>
     </div>
     <template v-for="loader in 12">
       <Skeleton height="2rem" class="mb-2">{{loader}}</Skeleton>
     </template>
   </div>
   <template v-else>
-    <DataTable :value="carouselStore.carousel" tableStyle="min-width: 50rem">
+    <DataTable :value="menuStore.menu" tableStyle="min-width: 50rem">
       <template #header>
         <div class="flex flex-wrap align-items-center justify-content-between gap-2">
-          <span class="text-xl text-900 font-bold">Carrusel de imágenes</span>
+          <span class="text-xl text-900 font-bold">Menú</span>
           <Button
             icon="pi pi-refresh"
             rounded
             raised
-            v-tooltip.top="`Actualizar imágenes en el carrusel`"
-            @click="carouselStore.getCarousel(true)"
-            :disabled="carouselStore.carousel.length === 0"
+            v-tooltip.top="`Actualizar menú`"
+            @click="menuStore.getMenu(true)"
+            :disabled="menuStore.menu.length === 0"
           />
           <Button
             icon="pi pi-plus"
             rounded
             raised
-            v-tooltip.top="`Agregar imagen en el carrusel`"
+            v-tooltip.top="`Agregar menú`"
             severity="info"
-            @click="manageModalCarousel('add', 'carousel')"
+            @click="manageModalMenu('add', 'menu')"
           />
         </div>
       </template>
-      <Column field="title" header="Nombre"></Column>
-      <Column header="Imagen">
-        <template #body="slotProps">
-          <img
-            :src="slotProps.data.image"
-            :alt="slotProps.data.title"
-            class="w-6rem border-round w-[20rem]"
-          />
-        </template>
-      </Column>
-      <Column header="Link">
+      <Column field="title" header="Titulo"></Column>
+      <Column field="name" header="Nombre"></Column>
+      <Column field="link" header="Link">
         <template #body="slotProps">
           <a
             v-if="slotProps.data.link"
@@ -130,24 +120,32 @@ onMounted(() => {
           </a>
         </template>
       </Column>
+      <Column field="order" header="Orden"></Column>
+      <Column field="isVisible" header="Visible">
+        <template #body="slotProps">
+          <span v-if="slotProps.data.isVisible" class="text-green-500">Si</span>
+          <span v-else class="text-red-500">No</span>
+        </template>
+      </Column>
       <Column header="Acción">
         <template #body="slotProps">
           <div class="flex gap-2">
             <Button
               icon="pi pi-pencil"
               class="p-button-rounded p-button-success p-button-outlined"
-              v-tooltip.top="`Editar imagen del carrusel`"
+              v-tooltip.top="`Editar menú`"
               severity="info" rounded outlined
-              aria-label="Editar imagen del carrusel"
-              @click="manageModalCarousel('edit', 'carousel', slotProps.data)"
+              aria-label="Editar menú"
+              @click="manageModalMenu('edit', 'menu', slotProps.data)"
             />
             <Button
               icon="pi pi-trash"
               class="p-button-rounded p-button-danger p-button-outlined"
-              v-tooltip.top="`Eliminar imagen del carrusel`"
+              v-tooltip.top="`Eliminar menú`"
               severity="danger" text rounded
-              aria-label="Eliminar imagen del carrusel"
-              @click="confirmDeleteCarousel(slotProps.data.id, slotProps.data.image)"
+              aria-label="`Eliminar menú"
+              @click="confirmDeleteMenu(slotProps.data.id)"
+              :disabled="slotProps.data.readOnly"
             />
           </div>
         </template>