+++ src/views/Admin.vue
@@ -4,6 +4,7 @@ import { useHead } from '@unhead/vue'
 
 const MpAdminCatalogs = defineAsyncComponent(/* webpackChunkName: "mpAdminCatalogs" */() => import('@/components/admin/MpAdminCatalogs.vue'))
 const MpAdminCarousel = defineAsyncComponent(/* webpackChunkName: "mpAdminCarousel" */() => import('@/components/admin/MpAdminCarousel.vue'))
+const MpAdminMenu = defineAsyncComponent(/* webpackChunkName: "mpAdminMenu" */() => import('@/components/admin/MpAdminMenu.vue'))
 
 useHead({
   title: '🔐 Admin | Megapromocionales',
@@ -14,8 +15,8 @@ useHead({
   <ConfirmDialog />
   <div class="container mx-auto lg:py-5">
     <TabView>
-      <TabPanel header="Menú" :disabled="true">
-        <p>Sin información</p>
+      <TabPanel header="Menú">
+        <MpAdminMenu />
       </TabPanel>
       <TabPanel header="Carrusel de Imágenes">
         <MpAdminCarousel />